# backend/docs/swagger.yaml
definitions:
  dto.AddRetakeStudentRequest:
    properties:
      debt_id:
        type: string
      student_id:
        type: string
    type: object
  dto.AddRetakeTeacherRequest:
    properties:
      teacher_id:
        type: string
    type: object
...
  dto.CreateRetakeRequest:
    properties:
      building:
        type: string
      discipline_id:
        type: string
      duration_minutes:
        type: integer
      kind:
        type: string
      notes:
        type: string
      room:
        type: string
      scheduled_at:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.GradeRetakeStudentRequest:
    properties:
      grade:
        type: integer
    type: object
...
  dto.RetakeParticipantResponse:
    properties:
      created_at:
        type: string
      debt_id:
        type: string
      grade:
        type: integer
      graded_at:
        type: string
      graded_by:
        type: string
      id:
        type: string
      kind:
        description: student / teacher / commission_member
        type: string
      retake_id:
        type: string
      user_id:
        type: string
    type: object
  dto.RetakeResponse:
    properties:
      building:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      discipline_id:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      kind:
        description: regular / commission
        type: string
      min_teachers:
        type: integer
      notes:
        type: string
      room:
        type: string
      scheduled_at:
        type: string
      status:
        description: scheduled / in_progress / completed / cancelled
        type: string
      updated_at:
        type: string
    type: object
  dto.RetakesListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.RetakeResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
  dto.UpdateRetakeRequest:
    properties:
      building:
        type: string
      duration_minutes:
        type: integer
      notes:
        type: string
      room:
        type: string
      scheduled_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/retakes:
    get:
      parameters:
      - description: 'Фильтр: scheduled|in_progress|completed|cancelled'
        in: query
        name: status
        type: string
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RetakesListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Все пересдачи (деканат)
      tags:
      - retakes
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пересдачи
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRetakeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RetakeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать пересдачу (деканат)
      tags:
      - retakes
  /api/retakes/{id}:
    get:
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RetakeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Пересдача по ID
      tags:
      - retakes
    patch:
      consumes:
      - application/json
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      - description: Поля для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRetakeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RetakeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить расписание пересдачи
      tags:
      - retakes
  /api/retakes/{id}/cancel:
    post:
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отменить пересдачу
      tags:
      - retakes
  /api/retakes/{id}/complete:
    post:
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Завершить пересдачу вручную
      tags:
      - retakes
  /api/retakes/{id}/participants:
    get:
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RetakeParticipantResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Участники пересдачи
      tags:
      - retakes
  /api/retakes/{id}/start:
    post:
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "409":
          description: Неверный статус
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Начать пересдачу вручную
      tags:
      - retakes
  /api/retakes/{id}/students:
    post:
      consumes:
      - application/json
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      - description: StudentID + DebtID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddRetakeStudentRequest'
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить студента на пересдачу
      tags:
      - retakes
  /api/retakes/{id}/students/{user_id}:
    delete:
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      - description: UUID студента
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Убрать студента с пересдачи
      tags:
      - retakes
  /api/retakes/{id}/students/{user_id}/grade:
    patch:
      consumes:
      - application/json
      description: Атомарно закрывает связанный долг студента.
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      - description: UUID студента
        in: path
        name: user_id
        required: true
        type: string
      - description: Оценка
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.GradeRetakeStudentRequest'
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Выставить оценку студенту на пересдаче
      tags:
      - retakes
  /api/retakes/{id}/teachers:
    post:
      consumes:
      - application/json
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      - description: TeacherID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddRetakeTeacherRequest'
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить преподавателя на пересдачу
      tags:
      - retakes
  /api/retakes/{id}/teachers/{user_id}:
    delete:
      parameters:
      - description: UUID пересдачи
        in: path
        name: id
        required: true
        type: string
      - description: UUID преподавателя
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Убрать преподавателя с пересдачи
      tags:
      - retakes
  /api/retakes/my:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RetakeResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Мои пересдачи (студент / преподаватель)
      tags:
      - retakes
...
swagger: "2.0"